+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Updated node to v12
 * Fixes regarding migration zapier-platform-core from v9 to v10, see [changelog](https://github.com/zapier/zapier-platform/blob/master/CHANGELOG.md#1000).
 * Migrated from travis into github actions
+* Fixed warning regarding omitted items from dataset
 
 ## 2.1.8 / 2020-02-05